pyproject.toml
[project]
name = "mvld"
version = "0.1.0"
description = "Manim Visual Logic Drift Post-Training System"
dependencies = [
    "manim>=0.18.1",
    "torch",
    "torchvision",
    "clip @ git+https://github.com/openai/CLIP.git",
    "typer[all]",
    "rich",
    "datasets",
    "transformers",
    "jsonlines",
    "pillow",
    "numpy",
]

[project.scripts]
mvld = "mvld.cli.main:app"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
